+++ src/constants.ts
@@ -1,8 +1,8 @@
-export const DAY_LABELS = ["월", "화", "수", "목", "금", "토"] as const;
+export const DAY_LABELS = ['월', '화', '수', '목', '금', '토'] as const;
 
 export const CellSize = {
-  WIDTH: 80,
-  HEIGHT: 30,
+	WIDTH: 80,
+	HEIGHT: 30,
 };
 
 export const 초 = 1000;
